feat: simplify makefiles and CI
this commit largely removes the redundency of writing each command once for each shim
this repo supports. It uses foreach syntax to iterate over each item in the shims
variable and execute a command for each one.

This also applies to the CI. This commit adds a new docker-build-push action that
builds and pushes the images to ghcr, abstrating the repeated build and push steps
into this reusable workflow.

It also adds spin-inbound-redis and spin-outbound-redis images to the release pipeline

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>