fix: make the spin shim work
this commits deletes the runtime config path which isn't being used
in the spin trigger. It also removes some commented lines of the stdio
becuase now in the cgroup all the ios are handled by the prepare_stdio
function and spin can just use the default stdio wiring up wasmtime.
As a consequence of it, we can also remove podio.rs

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>