feat: integrate youki to the spin shim (#119)
* feat: integrate youki to the spin shim

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>

* fix: run clippy and rustfmt

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>

* fix: make the spin shim work

this commits deletes the runtime config path which isn't being used
in the spin trigger. It also removes some commented lines of the stdio
becuase now in the cgroup all the ios are handled by the prepare_stdio
function and spin can just use the default stdio wiring up wasmtime.
As a consequence of it, we can also remove podio.rs

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>

* fix: remove command from the workloads

This commits removes the command from workloads yamls. This is because
now the spin image has an explicit entrypoint pointing to a
executable wasm module, which is required by the libcontainer.

I am expecting that this will go away with futuer iteration of the
libcontainer

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>

* rustfmt

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>

* feat: add more observability to the CI cluster

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>

* fix: add a clean up step at the end of the build CI

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>

---------

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>