feat(*): remove spin shim from the repo
this commit removes the spin shim from the repo
since we are moving the spin shim to SpinKube
organization all-together.

It also removes spin-related images and cluster
configuration. In the future release, there will
not be a spin shim baked into the k3d cluster
and no published spin-related images

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>